Small correction to erroneous prototype of hashCode (relevant for 64-bits).

+++ sources/gkmod/kgb.cpp
@@ -390,7 +390,7 @@ void KGB::fillBruhat()
 namespace kgb {
   namespace {
 
-size_t TI_Entry::hashCode(unsigned int modulus) const
+size_t TI_Entry::hashCode(size_t modulus) const
 {
   unsigned int hash=0;
   for (size_t i=constants::RANK_MAX; i-->0; )
@@ -398,7 +398,7 @@ size_t TI_Entry::hashCode(unsigned int modulus) const
   return hash & modulus-1;
 }
 
-size_t TE_Entry::hashCode(unsigned int modulus) const
+size_t TE_Entry::hashCode(size_t modulus) const
 {
   unsigned int hash=0;
   for (size_t i=constants::RANK_MAX; i-->0; )
